user list refactor (names and structure)
