--- src/client/components/UserList/ButtonUserListExport/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ButtonUserListExport'
