+++ src/client/components/UserList/UserListButtonExport/index.ts
@@ -1,0 +1,1 @@
+export { default } from './UserListButtonExport'
