+++ src/client/components/UserList/UserListElement/UserField/UserField.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+import { User } from '@meta/user'
+
+const UserField: React.FC<{ user: User; field: keyof User }> = ({ user, field }) => (
+  <td className="user-list__cell">
+    <div className="user-list__cell--read-only">{user[field] ? (user[field] as string) : '\xA0'}</div>
+  </td>
+)
+
+export default UserField
