+++ src/client/components/UserList/UserListElement/UserRoleField/index.ts
@@ -1,0 +1,1 @@
+export { default } from './UserRoleField'
