+++ src/client/components/UserList/UserListElement/index.ts
@@ -1,0 +1,1 @@
+export { default } from './UserListElement'
