+++ src/client/components/UserList/UserListHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './UserListHeader'
